requirements.txt
google-cloud == 0.34.0
google-auth == 2.9.1
pandas == 1.4.3
regex == 2022.7.25
numpy == 1.23.1